+++ src/actions/types.ts
@@ -67,6 +67,7 @@ export type ActionName =
   | "changeProjectName"
   | "changeExportBackground"
   | "changeExportEmbedScene"
+  | "changeExportScale"
   | "saveToActiveFile"
   | "saveFileToDisk"
   | "loadScene"